Rename Sdk directory
