+++ Sdk.sln
@@ -1,7 +1,7 @@
 ﻿
 Microsoft Visual Studio Solution File, Format Version 12.00
 # 
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "TrueLayerSdk", "src\TrueLayerSdk\TrueLayerSdk.csproj", "{3CAD6ACA-344B-4FC3-965E-8C116759B50A}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "TrueLayerSdk", "src\Sdk\Sdk.csproj", "{3CAD6ACA-344B-4FC3-965E-8C116759B50A}"
 EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "src", "src", "{A28C8509-D982-426E-A9DC-83A169621955}"
 EndProject
