+++ docs/index.md
@@ -1,8 +1,8 @@
 # TrueLayer SDK
 
-[![NuGet](https://img.shields.io/nuget/v/TrueLayerSDK.svg)](https://www.nuget.org/packages/TrueLayerSDK)
-[![NuGet](https://img.shields.io/nuget/vpre/TrueLayerSDK?label=Pre-release)](https://www.nuget.org/packages/TrueLayerSDK)
-[![NuGet](https://img.shields.io/nuget/dt/TrueLayerSDK.svg)](https://www.nuget.org/packages/TrueLayerSDK)
+[![NuGet](https://img.shields.io/nuget/v/TrueLayerSdk.svg)](https://www.nuget.org/packages/TrueLayerSdk)
+[![NuGet](https://img.shields.io/nuget/vpre/TrueLayerSdk?label=Pre-release)](https://www.nuget.org/packages/TrueLayerSdk)
+[![NuGet](https://img.shields.io/nuget/dt/TrueLayerSdk.svg)](https://www.nuget.org/packages/TrueLayerSdk)
 [![License](https://img.shields.io/:license-mit-blue.svg)](https://benfoster.mit-license.org/)
 
 ## Getting Started
