+++ src/Sdk/AssemblyInfo.cs
@@ -1,0 +1,2 @@
+using System.Runtime.CompilerServices;
+[assembly: InternalsVisibleToAttribute("TrueLayer.Sdk.Tests")]
