+++ src/Sdk/Auth/AuthClient.cs
