+++ src/Sdk/Auth/ExchangeCodeResponse.cs
