+++ src/Sdk/Auth/GetAuthUriRequest.cs
