+++ src/Sdk/Auth/GetAuthUriResponse.cs
