+++ src/Sdk/Auth/IAuthClient.cs
