+++ src/Sdk/Auth/Model/GetPaymentTokenRequest.cs
