+++ src/Sdk/Auth/Model/GetPaymentTokenResponse.cs
