+++ src/Sdk/ErrorDetails.cs
