+++ src/Sdk/ErrorResponse.cs
