+++ src/Sdk/Serialization/SnakeCaseNamingPolicy.cs
