+++ src/Sdk/Serialization/StringExtensions.cs
