--- src/TrueLayerSdk/AssemblyInfo.cs
@@ -1,2 +1,0 @@
-using System.Runtime.CompilerServices;
-[assembly: InternalsVisibleToAttribute("TrueLayer.Sdk.Tests")]
